requirements.txt
adagio==0.2.6
aiohappyeyeballs==2.6.1
aiohttp==3.13.2
aiosignal==1.4.0
anyio==4.12.0
attrs==25.4.0
catboost==1.2.8
certifi==2025.11.12
cftime==1.6.5
charset-normalizer==3.4.4
click==8.3.1
cloudpickle==3.1.2
cmdstanpy==1.3.0
contourpy==1.3.3
coreforecast==0.0.16
cycler==0.12.1
filelock==3.20.0
fonttools==4.61.0
frozenlist==1.8.0
fsspec==2025.12.0
fugue==0.9.3
graphviz==0.21
h11==0.16.0
hf-xet==1.2.0
holidays==0.86
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==1.2.1
idna==3.11
importlib-resources==6.5.2
jinja2==3.1.6
joblib==1.5.2
kiwisolver==1.4.9
lightgbm==4.6.0
lightning-utilities==0.15.2
llvmlite==0.45.1
markupsafe==3.0.3
matplotlib==3.10.7
mpmath==1.3.0
multidict==6.7.0
narwhals==2.13.0
netcdf4==1.7.3
networkx==3.6.1
nfoursid==1.0.2
numba==0.62.1
numpy==2.3.5
packaging==25.0
pandas==2.3.3
patsy==1.0.2
pillow==12.0.0
plotly==6.5.0
propcache==0.4.1
prophet==1.2.1
protobuf==6.33.2
pyarrow==22.0.0
pyod==2.0.6
pyparsing==3.2.5
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.2
pytz==2025.2
pyyaml==6.0.3
requests==2.32.5
safetensors==0.7.0
scikit-learn==1.7.2
scipy==1.16.3
setuptools==80.9.0
shap==0.50.0
shellingham==1.5.4
six==1.17.0
slicer==0.0.8
stanio==0.5.1
statsforecast==2.0.1
statsmodels==0.14.5
sympy==1.14.0
tensorboardx==2.6.4
threadpoolctl==3.6.0
torch==2.9.1
torchmetrics==1.8.2
tqdm==4.67.1
triad==1.0.0
typer-slim==0.20.0
typing-extensions==4.15.0
tzdata==2025.2
u8darts==0.39.0
urllib3==2.5.0
utilsforecast==0.2.15
xarray==2025.11.0
xgboost==3.1.2
yarl==1.22.0
